Squashed commit of the following:
commit 85a42476adde39d32d607de33202736712fa50fb
Merge: f58c375fc 4241ce453
Author: Divyanshu Agrawal <agrawal.divyanshu@outlook.com>
Date:   Wed Oct 21 12:48:11 2020 +0530

    Merge branch 'master' of https://github.com/arnav-ag/cph into arnav-ag-master

commit 4241ce453cb33fedf10876dca06351ef18002e1f
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Tue Oct 20 15:47:50 2020 +0800

    Fix conflict issues

commit 41a45a2e372d98f557473c7942c401643c789042
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Tue Oct 20 15:43:53 2020 +0800

    Fix conflict issues

commit f0c7853bc318688a9fbed1167525a514acaf90de
Merge: 943450e55 8cfe2d5b5
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Fri Oct 16 18:06:06 2020 +0800

    Merge branch 'master' of github.com:arnav-ag/cph into master

commit 943450e55eca03f11016519c196831619cd13604
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Fri Oct 16 17:52:11 2020 +0800

    Fix conflict issues

commit 8cfe2d5b5fe824344d56ea970f7fa404ad392f8d
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Fri Oct 16 17:56:42 2020 +0800

    Fix conflict issues

commit a1a4afac51d4dd7904a4efb96b8c7478573af84e
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Fri Oct 16 17:55:52 2020 +0800

    Fix conflict issues

commit fe84bec1524e113604a94986d4e3cfbd2ff983b6
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Fri Oct 16 17:52:11 2020 +0800

    Fix conflict issues

commit 6478fc55bcb49db77253890955ca0bb810cbc7d3
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Thu Oct 15 21:25:43 2020 +0800

    Fix checking issues

commit 027893540a04c0743f6998c659f3c2201638230d
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Thu Oct 15 21:15:25 2020 +0800

    Fix npm install issue

commit 8db23e74c0bb9233094e9d615d62d95c55270e3b
Merge: 68b82d308 abc53e348
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Thu Oct 15 19:48:12 2020 +0800

    Fix resolve conflicts

commit 68b82d30851ae540810180e1fda4d75c3074b8ac
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Thu Oct 15 19:45:05 2020 +0800

    Fix resolve conflicts

commit b1c5228af01562bf20ec64c7b7054bdce4c54b16
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Thu Oct 15 19:35:45 2020 +0800

    Fix resolve conflicts

commit ff63f18997b10140663c21c1a44d27ec5f48529a
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Thu Oct 15 19:24:16 2020 +0800

    Fix resolve conflicts

commit 952c0250aaaaec6ba53bfedc25257fa3d6158087
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Thu Oct 15 19:20:34 2020 +0800

    Fix resolve conflicts

commit 7353095e77699587948d33c5c8776248b44d0747
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Thu Oct 15 18:56:42 2020 +0800

    Changed to use python3, removed unnecessary imports

commit fc4ce76f75be883ef7dd8cf426a1b9986097c289
Merge: 59099e176 ad4de4f29
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Fri Aug 28 19:13:59 2020 +0800

    Fix resolve conflicts

commit 59099e1764dcfdb2c3b40893632f531b8c1ed7e3
Merge: 62310b78f 2cc9cfb97
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Wed Aug 19 14:07:30 2020 +0800

    Merge branch 'master' of github.com:arnav-ag/cph

commit 62310b78f6cd990318c3e1d5d0c88f0f4aa476ef
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Wed Aug 19 14:07:00 2020 +0800

    Added 'Submitting...' feedback, sends submit-finished message to webview

commit 2cc9cfb975ea83bb99ee0bbc1be945c941b3cb03
Author: Arnav Aggarwal <65700476+arnav-ag@users.noreply.github.com>
Date:   Wed Aug 19 12:15:06 2020 +0800

    Delete kattis-competitive-programming-helper-4.3.6.vsix

commit c70870d40ba78d9f556e1a082d09c0aa48f969a3
Author: Arnav Aggarwal <65700476+arnav-ag@users.noreply.github.com>
Date:   Wed Aug 19 12:13:53 2020 +0800

    Update package.json

commit aa52577f2d2f507d569a29ece83e4ca76741559e
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Wed Aug 19 12:08:37 2020 +0800

    Fixed working on Windows with proper error messages and handling.

commit a650bb5bcc6b7a4c2bd0182fe741e6c08f2055e2
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Tue Aug 18 23:56:38 2020 +0800

    Added vsix file for download

commit d9b4dde54f1f4d30c8af061babd0575edca2123b
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Tue Aug 18 20:28:28 2020 +0800

    Fix linting issues

commit d595e55445585cff73649f24450f64e8b28715ef
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Tue Aug 18 20:22:30 2020 +0800

    Fixed wrong style of apostrophy

commit 15ddd812139736924b5f2af3884d7a9639b53f16
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Tue Aug 18 20:04:02 2020 +0800

    Changed such that filenames use kattis problem id for easier submission

commit a882c69d28b0deb48b51caaa3dd0c8e1cc7e3998
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Tue Aug 18 18:57:59 2020 +0800

    Update version

commit 6c5442702ff8ae05f08e9aa0571da9f745b17111
Author: arnav-ag <arnav.ctf@gmail.com>
Date:   Tue Aug 18 18:46:31 2020 +0800

    Fixed submit for Kattis.

commit 75eb836f2c502d4cf0f2dcbfbe781bd6ad0f06bf
Author: xX0206Xx <indiarnav@gmail.com>
Date:   Tue Aug 18 17:14:05 2020 +0800

    what am i doing pls help

commit b204f171376acf35f4a7e3f43a6abe3b0a6da815
Author: xX0206Xx <indiarnav@gmail.com>
Date:   Tue Aug 18 16:15:50 2020 +0800

    Adding support for kattis

+++ src/companion.ts
@@ -4,7 +4,7 @@ import { Problem, CphSubmitResponse, CphEmptyResponse } from './types';
 import { saveProblem } from './parser';
 import * as vscode from 'vscode';
 import path from 'path';
-import { writeFileSync, readFileSync } from 'fs';
+import { writeFileSync, readFileSync, existsSync } from 'fs';
 import {
     startWebVeiwIfNotActive,
     setBaseWebViewHTML,
@@ -17,9 +17,57 @@ import {
     useShortCodeForcesName,
 } from './preferences';
 import { getProblemName } from './submit';
+import { spawn } from 'child_process';
 
 const emptyResponse: CphEmptyResponse = { empty: true };
 let savedResponse: CphEmptyResponse | CphSubmitResponse = emptyResponse;
+
+export const submitKattisProblem = (problem: Problem) => {
+    const srcPath = problem.srcPath;
+    const homedir = require('os').homedir();
+    let submitPath = `${homedir}/.kattis/submit.py`;
+    //vscode.window.showInformationMessage(homedir);
+    if (process.platform == 'win32') {
+        if (
+            !existsSync(`${homedir}\\.kattis\\.kattisrc`) ||
+            !existsSync(`${homedir}\\.kattis\\submit.py`)
+        ) {
+            vscode.window.showErrorMessage(
+                `Please ensure .kattisrc and submit.py are present in ${homedir}\\.kattis\\submit.py`,
+            );
+            return;
+        } else {
+            submitPath = `${homedir}\\.kattis\\submit.py`;
+        }
+    } else {
+        if (
+            !existsSync(`${homedir}/.kattis/.kattisrc`) ||
+            !existsSync(`${homedir}/.kattis/submit.py`)
+        ) {
+            vscode.window.showErrorMessage(
+                `Please ensure .kattisrc and submit.py are present in ${homedir}/.kattis/submit.py`,
+            );
+            return;
+        } else {
+            submitPath = `${homedir}/.kattis/submit.py`;
+        }
+    }
+    const pyshell = spawn('python', [submitPath, '-f', srcPath]);
+
+    //tells the python script to open submission window in new tab
+    pyshell.stdin.setDefaultEncoding('utf-8');
+    pyshell.stdin.write('Y\n');
+    pyshell.stdin.end();
+
+    pyshell.stdout.on('data', function (data) {
+        console.log(data.toString());
+        extensionToWebWiewMessage({ command: 'submit-finished' });
+    });
+    pyshell.stderr.on('data', function (data) {
+        console.log(data.tostring());
+        vscode.window.showErrorMessage(data);
+    });
+};
 
 /** Stores a response to be submitted to CF page soon. */
 export const storeSubmitProblem = (problem: Problem) => {
@@ -122,8 +170,9 @@ const handleNewProblem = async (problem: Problem) => {
         ...testcase,
         id: randomId(),
     }));
-
-    writeFileSync(srcPath, '');
+    if (!existsSync(srcPath)) {
+        writeFileSync(srcPath, '');
+    }
     saveProblem(srcPath, problem);
     const doc = await vscode.workspace.openTextDocument(srcPath);
     await vscode.window.showTextDocument(doc, vscode.ViewColumn.One);
